// fem-vision-studio/design-mockup/src/lib/types/steel/ForcePoint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InternalForces } from "./InternalForces";

/**
 * One sample point along a beam: position (mm from start) + governing
 * internal forces at that location for some load combination.
 */
export type ForcePoint = { combination_id: number, position_mm: number, forces: InternalForces, };